pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kokoro-bot"
version = "0.1.0"
description = "The Vibe Architect — Audio intelligence that reveals the magic hidden in conversations."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Rodrigo"},
]
dependencies = [
    "deepgram-sdk>=3.0.0",
    "python-dotenv>=1.0.0",
    "matplotlib>=3.8.0",
    "discord.py[voice]>=2.3.0",
    "discord-ext-voice-recv>=0.5.0",
    "PyNaCl>=1.5.0",
    "aiofiles>=23.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
kokoro = "kokoro.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
